soc/rockchip: add handler for usb-uart functionality
Some Rockchip SoCs provide the possibility to use a usb-phy as passthru for
the debug uart, making it possible to get console output without needing to
open the device.

This patch adds an early_initcall to enable this functionality conditionally
and also disables the corresponding usb controller in the devicetree.