clk: rockchip: add support for phase inverters
Most Rockchip socs have optional phase inverters connected to some
clocks that move the clock-phase by 180 degrees.
While not having any actual influence on the rate, the inverter
provides its own simple recalc_rate callback as relying on the
fallback in the framework is for "lazy developers" according
to the documentation in the code.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>