ARM: dts: rockchip: add touchscreen node for rk3288-veyron-minnie
According to Dmitry Torokhov shipping Minnie devices use elan,ekth3500
as touchscreen. So add the necessary node and properties to drive it.
There may be (preproduction) devices around that use an atmel,atmel_mxt_ts
sharing the same pins, but hopefully they do not turn up as production
devices too. If this happens we'll need to reevaluate this.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>