clk: defer clk_gets on orphan clocks
Orphan clocks or children of orphan clocks don't have rate information at
all and can produce strange results if they're allowed to be used and the
parent becomes available later on.

This change defers __clk_create_clk() calls on orphan clocks in all
regular cases.

One special case that gets handled, is accessing such orphan clocks when
handling assigned-clocks configurations. In the boot-defaults it may be
the case that a clock is connected to an orphan parent which then should
be reparented to an actually usable clock using assigned-clock-parents.

In this case even orphaned clocks should be handled correctly, but only
for the set-parent case.

Signed-off-by: Heiko Stuebner <heiko.stuebner@collabora.com>